feat: eager-load user data when fetching mementos, uplift model serialisation

+++ controller/memento_controller.go
@@ -23,7 +23,7 @@ func GetMementos(w http.ResponseWriter, r *http.Request) {
 	}
 
 	var memento = models.Memento{
-		UserID: userid,
+		UserID: uint(userid),
 	}
 
 	mementos, err := memento.GetMementosByUserId(context.Context.DB)
